requirements.txt
alembic==1.0.10
autopep8==1.4.4
blinker==1.4
Click==7.0
Flask==1.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
trash-cli==0.17.1.14
Werkzeug==0.15.4
WTForms==2.2.1
